Honours_code/requirements.txt
Flask 
azure-cognitiveservices-search-imagesearch  
pillow 
cognitive_face
pymongo